requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0,<1.5.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-jose[cryptography]>=3.3.0,<3.4.0
python-multipart>=0.0.5,<0.0.6
email-validator>=1.1.3,<1.2.0
python-dotenv>=0.19.0,<0.20.0
pymysql>=1.0.2,<1.1.0
cryptography>=35.0.0,<36.0.0
python-dateutil>=2.8.2
pytest>=7.0.0
pytest-asyncio>=0.18.0
httpx>=0.23.0
alembic>=1.7.0
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
mypy>=0.910
redis>=4.3.0,<4.4.0
python-memcached>=1.59,<1.60
sentry-sdk>=1.15.0,<1.16.0
prometheus-client>=0.15.0,<0.16.0
starlette-exporter>=0.3.0,<0.4.0 